Rename elt cli to cli_utils
Interim change to keep the current pipelines functioning whilst a new CLI is put in place for #321
